stream: rename StallingManager into simple getStalledEvents function
