+++ src/core/stream/stream_directfile.ts
@@ -41,8 +41,8 @@ import { IKeySystemOption } from "../eme/types";
 import createEMEManager from "./create_eme_manager";
 import EVENTS from "./events_generators";
 import { IInitialTimeOptions } from "./get_initial_time";
+import getStalledEvents from "./get_stalled_events";
 import seekAndLoadOnMediaEvents from "./initial_seek_and_play";
-import StallingManager from "./stalling_manager";
 import throwOnMediaError from "./throw_on_media_error";
 import {
   ISpeedChangedEvent,
@@ -156,7 +156,7 @@ export default function StreamDirectFile({
 
   // Create Stalling Manager, an observable which will try to get out of
   // various infinite stalling issues
-  const stallingManager$ = StallingManager(mediaElement, clock$)
+  const stalled$ = getStalledEvents(mediaElement, clock$)
     .pipe(map(EVENTS.stalled));
 
   // Manage "loaded" event and warn if autoplay is blocked on the current browser
@@ -183,7 +183,7 @@ export default function StreamDirectFile({
     emeManager$,
     mediaError$,
     playbackRate$,
-    stallingManager$,
+    stalled$,
     linkURL$
   );
 }
